Move remaining connection management logic into ConnectionManager
